chore: clean the code and not return error if no conversation
