--- pkg/component/application/instillapp/v0/main_test.go
@@ -1,2 +1,0 @@
-// TODO: chuang8511
-package instillapp
